Add new tests for category, brand and product

+++ MilaIrisWebParent/MilaIrisBackEnd/src/main/java/kz/milairis/admin/category/CategoryController.java
@@ -41,7 +41,7 @@ public class CategoryController {
 		CategoryPageInfo pageInfo = new CategoryPageInfo();
 		List<Category> listCategories = service.listByPage(pageInfo, pageNum, sortDir, keyword);
 
-		long startCount = (pageNum - 1) * CategoryService.ROOT_CATEGORIES_PER_PAGE + 1;
+		long startCount = (long) (pageNum - 1) * CategoryService.ROOT_CATEGORIES_PER_PAGE + 1;
 		long endCount = startCount + CategoryService.ROOT_CATEGORIES_PER_PAGE - 1;
 		if (endCount > pageInfo.getTotalElements()) {
 			endCount = pageInfo.getTotalElements();
